Updates to Projects Page and Navbar
Created the `Projects` page and it retrieves and displays all the projects and their information. I also had to put the `projectImages` folder into the projects folder in order for the source path would work. In addition to the projects page, I restyled the navbar; it may need some usability and accessibility improvements.