New Skills Table
Created a new Skills table component for the About page. The table retrieves data from the skillsData.ts file, processes it into an array with the skill names, level (proficiency), and years of experience, and then also creates a seperate array with the corresponding levels so that each skill's level will appear in a font-color that goes with their level; levels consist of "Beginner", "Intermediate", "Advanced", and "Expert". The skill data is then mapped into the corresponding table columns and rows.
Each column can be sorted in ascending order by clicking on the column headings. There is also a search box to query for a specific skill in the table. Rows can also be selected.